% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparCEV_function.R
\name{SparCEV}
\alias{SparCEV}
\title{Estimation of Cross-correlations between OTU abundances (or components of other
compositional variables) and a non-compositional variable.}
\usage{
SparCEV(
  OTU.abn,
  phenotype,
  pseudo_count = 1,
  var_min = 1e-05,
  iter = 20,
  t = NULL,
  B_t = 100,
  t_quant = 0.8,
  cores = 1,
  Find_m = TRUE,
  B_m = 100
)
}
\arguments{
\item{OTU.abn}{An OTU table with observarions as rows and OTUs columns
(or a similar compositional dataset).}

\item{phenotype}{Continuous non-compositional phenotypic variable.}

\item{pseudo_count}{The pseudo-count added to all read counts to avoid division by zero.}

\item{var_min}{In the event that a variance estimate of an OTU is negative,
it is replaced by this number.}

\item{iter}{The maximum number of iteration carried out.}

\item{t}{A cutoff below which a correlation is considered weak. Only
OTUs weakly correlated with the phenotypic variable are included in the next
iteration. When t = NULL, it is chosen via bootstrap.}

\item{B_t}{The number of bootstraps to use when selecting t. Only used when t = NULL.}

\item{t_quant}{The quantile used under bootstrap for selecting t. Only used whern t = NULL.}

\item{cores}{The number of cores used for selecting t and m.}

\item{Find_m}{Toggles whether or not the permutation threshold, m, should be
calculated and returned. All correlations above m are likely to be real signals.}

\item{B_m}{The number of bootstraps to use when calculating m. Only used when Find_m = TRUE.}
}
\value{
A list containing the following: cor: a vector containing the
correlations between the OTU abundances and the phenotypic variable, Final_idx:
A logical vector indicating whether or not an OTU was included in the final iteration,
t: The value of t (either user specified or returned by bootstrap), m: the value of
m (NULL if Find_m = FALSE), Correlated: A logical vector indicating if an OTU's estimated
absolute correlation coefficient is above m.
}
\description{
Estimation of Cross-correlations between OTU abundances (or components of other
compositional variables) and a non-compositional variable.
}
