% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparCEV_function.R
\name{SparCEV_base}
\alias{SparCEV_base}
\title{Estimation of Cross-correlations between OTU abundances (or components of other
compositinal datasets) and a non-compositional varible. This method is appropriate
when the average covariance is close to zero.}
\usage{
SparCEV_base(
  OTU.abn,
  phenotype,
  pseudo_count = 1,
  var_min = 1e-05,
  cores = 1,
  Find_m = TRUE,
  B_m = 100
)
}
\arguments{
\item{OTU.abn}{An OTU table with observarions as rows and OTUs columns
(or a similar compositional dataset).}

\item{phenotype}{Continuous non-compositional phenotypic variable.}

\item{pseudo_count}{The pseudo-count added to all read counts to avoid division by zero.}

\item{var_min}{In the event that a variance estimate of an OTU is negative,
it is replaced by this number.}

\item{cores}{The number of cores used for finding m.}

\item{Find_m}{Toggles whether or not the permutation threshold, m, should be
calculated and returned. All correlations above m are likely to be real signals.}

\item{B_m}{The number of bootstraps to use when calculating m. Only used when Find_m = TRUE.}
}
\value{
A vector containing the correlations between the OTU abundances and the phenotypic variable.
}
\description{
Estimation of Cross-correlations between OTU abundances (or components of other
compositinal datasets) and a non-compositional varible. This method is appropriate
when the average covariance is close to zero.
}
