% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparXCC_function.R
\name{SparXCC}
\alias{SparXCC}
\title{Estimation of Cross-correlations between OTU abundances (or components of other
compositional variables) and gene expressions (or components of other
compositional variables).}
\usage{
SparXCC(
  OTU.abn,
  gene.expr,
  pseudo_count = 1,
  var_min = 1e-05,
  iter = 10,
  t1 = NULL,
  t2 = NULL,
  B_t = 100,
  t1_quant = 0.8,
  t2_quant = 0.8,
  Find_m = TRUE,
  B_m = 100,
  cores = 1
)
}
\arguments{
\item{OTU.abn}{An OTU table with observarions as rows and OTUs columns
(or a similar compositional dataset).}

\item{gene.expr}{A table of gene expressions with observarions as rows and OTUs columns
(or a similar compositional dataset).}

\item{pseudo_count}{The pseudo-count added to all read counts to avoid division by zero.}

\item{var_min}{In the event that a variance estimate of an OTU is negative,
it is replaced by this number.}

\item{iter}{The maximum number of iteration carried out.}

\item{t1}{A cutoff to be used in the iterative procedure. Only
OTUs whose average absolute correlation with the gene expressions are included in the next
iteration. When t1 = NULL, it is chosen via bootstrap.}

\item{t2}{A cutoff to be used in the iterative procedure. Only
genes whose average absolute correlation with the OTU abundances are included in the next
iteration. When t2 = NULL, it is chosen via bootstrap.}

\item{B_t}{The number of bootstraps to use when selecting t1 and t2. Only used when t1 = NULL and/or t2 = NULL.}

\item{t1_quant}{The quantile used under bootstrap for selecting t1. Only used whern t1 = NULL.}

\item{t2_quant}{The quantile used under bootstrap for selecting t2. Only used whern t2 = NULL.}

\item{Find_m}{Toggles whether or not the permutation threshold, m, should be
calculated and returned. All correlations above m are likely to be real signals.}

\item{B_m}{The number of bootstraps to use when calculating m. Only used when Find_m = TRUE.}

\item{cores}{The number of cores used for selecting t and m.}
}
\value{
A list containing the following: cor: a matrix containing the
correlations between the OTU abundances and gene expressions, Final_idx_OTU:
A logical vector indicating whether or not an OTU was included in the final iteration,
Final_idx_gene: A logical vector indicating whether or not a gene was included in the final iteration,
t1: The value of t1 (either user specified or returned by bootstrap),
t2: The value of t2 (either user specified or returned by bootstrap),
m: the value of m (NULL if Find_m = FALSE), Correlated: A logical matrix
indicating if the absolute correlation coefficient estimate between and OTU
and a gene is above m.
}
\description{
Estimation of Cross-correlations between OTU abundances (or components of other
compositional variables) and gene expressions (or components of other
compositional variables).
}
