% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparXCC_function.R
\name{SparXCC_base}
\alias{SparXCC_base}
\title{Estimation of Cross-correlations between OTU abundances (or components of other
compositional variables) and gene expressions (or components of other
compositional variables). This method is appropriate
when the average covariance is close to zero.}
\usage{
SparXCC_base(
  OTU.abn,
  gene.expr,
  pseudo_count = 1,
  var_min = 1e-05,
  Find_m = TRUE,
  B_m = 100,
  cores = 1
)
}
\arguments{
\item{OTU.abn}{An OTU table with observarions as rows and OTUs columns
(or a similar compositional dataset).}

\item{gene.expr}{A table of gene expressions with observarions as rows and OTUs columns
(or a similar compositional dataset).}

\item{pseudo_count}{The pseudo-count added to all read counts to avoid division by zero.}

\item{var_min}{In the event that a variance estimate of an OTU is negative,
it is replaced by this number.}

\item{Find_m}{Toggles whether or not the permutation threshold, m, should be
calculated and returned. All correlations above m are likely to be real signals.}

\item{B_m}{The number of bootstraps to use when calculating m. Only used when Find_m = TRUE.}

\item{cores}{The number of cores used for finding m.}
}
\value{
A  matrix containing the correlations between the OTU abundances and gene expressions.
}
\description{
Estimation of Cross-correlations between OTU abundances (or components of other
compositional variables) and gene expressions (or components of other
compositional variables). This method is appropriate
when the average covariance is close to zero.
}
